Extract invite, part, join into membership
